requirements.txt
fastavro
kafka-python
python-dotenv
avro